+++ public/components/QnaMain.js
@@ -1,51 +1,7 @@
 import apis from '../apis.js';
 import {$} from '../utils/selector.js';
 import QnaMainNewQuestionModal from './QnaMainNewQuestionModal.js';
-
-function getAnswerTemplate(answers) {
-    return answers.reduce((html, {content, userId, date}) => {
-        return (
-            html +
-            `
-        <li class="answer-list" ">
-            <p class="answer-content">${content}</p>
-            <div class="answer-profile">
-                <span class="answer-writer">${userId} | </span>
-                <span class="answer-date">${date}</span>
-            </div>
-        </li>`
-        );
-    }, ``);
-}
-
-function getLoadingAnswerTpl() {
-    return `<li class="answer-list loading" ">
-        Loading.....
-     </li>`;
-}
-
-function getQnATemplate(data) {
-    return data.reduce((html, {title, question, id, matchedComments = []}) => {
-        return (
-            html +
-            ` <li class="qna" _questionId=${+id}>
-        <div class="qna-title">
-            <h2>${title}</h2>
-        </div>
-        <div class="question">
-            <p> ${question}</p>
-        </div>
-        <ul class="answer">${getAnswerTemplate(matchedComments)}</ul>
-        <div class="answer-form">
-            <form method="POST">
-                <textarea name="answer-content" class="answer-content-textarea" cols="30" rows="2" placeholder="새로운답변.."></textarea>
-            </form>
-            <button class="answer-submit">등록</button>
-        </div>
-    </li>`
-        );
-    }, ``);
-}
+import QnaMainQuestions from './QnaMainQuestions.js';
 
 export default function QnaMain({$el}) {
 
@@ -67,15 +23,19 @@ export default function QnaMain({$el}) {
         this.$el.innerHTML = `
             <div class="main-wrap">
                 <button class="new-question-btn" data-ref="new-question-open-btn">새로운질문</button>
-                <ul class="qna-wrap">
-                </ul>
-                
+                <div data-component="questions"></div>
                 <div data-component="new-question-modal"></div>
             </div>
         `;
 
-        $('.qna-wrap', this.$el).innerHTML = getQnATemplate(this.state.questions);
         this.components = {
+            questions: new QnaMainQuestions({
+                $el: $('[data-component="questions"]', this.$el),
+                props: {
+                    questions: this.state.questions,
+                },
+            }),
+
             newQuestionModal: new QnaMainNewQuestionModal({
                 $el: $('[data-component="new-question-modal"]', this.$el),
                 props: {
@@ -120,6 +80,8 @@ export default function QnaMain({$el}) {
      * @param question
      */
     const addNewQuestion = async ({title, question}) => {
+        //todo validate new question(title, question)
+
         await apis.createQuestion({title, question});
         openOrCloseNewQuestionModal(false);
